test: All child object to provide their own sort method
The TestBase class takes the sort parameter that is a
function called to transform uftrace output after the
execution of a test.

That said, the function that is passed as a string and the
resolving of the function is only done on TestBase itself
and not the current subclassed object.

This change allows a class that inherit TestBase of providing
its own sorting method that will be correctly resolved.